+++ R/run_sampler.R
@@ -152,7 +152,6 @@ run_sampler_omrf = function(spec) {
     lambda = p$lambda,
     target_acceptance = s$target_accept,
     max_tree_depth = s$nuts_max_depth,
-    num_leapfrogs = s$hmc_num_leapfrogs,
     pairwise_scaling_factors_nullable = p$pairwise_scaling_factors
   )
 
@@ -210,7 +209,6 @@ run_sampler_mixed_mrf = function(spec) {
     sampler_type = s$update_method,
     target_acceptance = s$target_accept,
     max_tree_depth = s$nuts_max_depth,
-    num_leapfrogs = s$hmc_num_leapfrogs,
     na_impute = m$na_impute,
     missing_index_discrete_nullable = m$missing_index_discrete,
     missing_index_continuous_nullable = m$missing_index_continuous
@@ -268,7 +266,6 @@ run_sampler_compare = function(spec) {
     nThreads = s$cores,
     seed = s$seed,
     update_method = s$update_method,
-    hmc_num_leapfrogs = s$hmc_num_leapfrogs,
     progress_type = s$progress_type,
     progress_callback = s$progress_callback
   )